+++ src/internal/data.ts
@@ -1,12 +1,22 @@
 import { Model, Schema, SchemaDefinition } from 'mongoose';
 
-import { EmptyVoidFn, VirtualOptions, AnyParamConstructor } from '../types';
+import { VirtualOptions, AnyParamConstructor } from '../types';
 
-export interface HooksPrePost {
-  pre: Map<string | RegExp, (error?: Error) => void>;
-  post: Map<string | RegExp, EmptyVoidFn>;
+export interface IPreHook {
+  method: string | RegExp;
+  func(error?: Error): void;
 }
-export interface PluginMap {
+
+export interface IPostHook {
+  method: string | RegExp;
+  func(): void;
+}
+
+export interface IHooks {
+  pre: IPreHook[];
+  post: IPostHook[];
+}
+export interface IPluginMap {
   mongoosePlugin(schema: Schema<any>, options: object): void;
   options: object;
 }
@@ -19,7 +29,7 @@ export enum WhatIsIt {
 }
 
 export interface DecoratedPropertyMetadata {
-  rawOptions: any;
+  origOptions: any;
   Type: AnyParamConstructor<any>;
   target: any;
   key: string;
@@ -34,8 +44,8 @@ export const models: Map<string, Model<any>> = new Map();
 /** Virtuals Map */
 export const virtuals: Map<string, Map<string, VirtualOptions>> = new Map();
 /** Hooks Map */
-export const hooks: Map<string, HooksPrePost> = new Map();
+export const hooks: Map<string, IHooks> = new Map();
 /** Plugins Map */
-export const plugins: Map<string, PluginMap[]> = new Map();
+export const plugins: Map<string, IPluginMap[]> = new Map();
 /** Constructors Map */
 export const constructors: Map<string, NewableFunction> = new Map();