+++ src/internal/utils.ts
@@ -1,9 +1,7 @@
 import * as mongoose from 'mongoose';
 
 import { isNullOrUndefined } from 'util';
-import { logger } from '../logSettings';
 import {
-  // IModelOptions,
   AnyParamConstructor,
   IModelOptions,
   PropOptionsWithNumberValidate,
@@ -243,20 +241,24 @@ export function mergeSchemaOptions<T, U extends AnyParamConstructor<T>>(value: m
  */
 export function getName<T, U extends AnyParamConstructor<T>>(cl: U) {
   const options: IModelOptions = Reflect.getMetadata(DecoratorKeys.ModelOptions, cl) || {};
+  const baseName = cl.name;
 
-  if (options.options && options.options.automaticName === false) {
-    if (typeof options.options.customName !== 'string') {
-      logger.warn('Option "automaticName" is set to "false" but no "customName" is provided! (Class: %s)', cl.name);
-    } else {
-      return options.options.customName;
-    }
+  if (options.options && options.options.automaticName) {
+    const suffix = (options.options ? options.options.customName : undefined) ||
+      (options.schemaOptions ? options.schemaOptions.collection : undefined);
+
+    return suffix ? `${baseName}_${suffix}` : baseName;
   }
 
-  const baseName = cl.name;
-  const suffix = (options.options ? options.options.customName : undefined) ||
-    (options.schemaOptions ? options.schemaOptions.collection : undefined);
+  if (options.options && typeof options.options.customName === 'string') {
+    if (options.options.customName.length <= 0) {
+      throw new TypeError(`"customName" must be a string AND at least one character ("${cl.name}")`);
+    }
 
-  return suffix ? `${baseName}_${suffix}` : baseName;
+    return options.options.customName;
+  }
+
+  return baseName;
 }
 
 /**